Refactor code
move iterator, rename packages
