--- src/global/goit/romashko/graphic_editor/my_collection/MyCollection.java
@@ -1,6 +1,0 @@
-package global.goit.romashko.graphic_editor.my_collection;
-
-public interface MyCollection<T> extends Iterable<T> {
-    void clear();
-    int size();
-}
