+++ src/global/goit/romashko/my_collection/MyCollection.java
@@ -1,0 +1,6 @@
+package global.goit.romashko.my_collection;
+
+public interface MyCollection<T> {
+    void clear();
+    int size();
+}
