+++ src/global/goit/romashko/my_collection/stack/MyArrayStack.java
@@ -1,20 +1,20 @@
-package global.goit.romashko.graphic_editor.my_collection.queue;
+package global.goit.romashko.my_collection.stack;
 
 import java.util.Iterator;
 import java.util.NoSuchElementException;
 import java.util.StringJoiner;
 
-public class MyArrayQueue<T> implements MyQueue<T> {
+public class MyArrayStack<T> implements MyStack<T> {
     private Object[] data;
     private int size;
 
-    public MyArrayQueue() {
-        assignEmptyQueue();
+    public MyArrayStack() {
+        assignEmptyStack();
     }
 
     @Override
     public void clear() {
-        assignEmptyQueue();
+        assignEmptyStack();
     }
 
     @Override
@@ -23,7 +23,7 @@ public class MyArrayQueue<T> implements MyQueue<T> {
     }
 
     @Override
-    public void add(T value) {
+    public void push(T value) {
         Object[] newData = new Object[size + 1];
         System.arraycopy(data,0,newData,0,size);
         newData[size] = value;
@@ -37,7 +37,7 @@ public class MyArrayQueue<T> implements MyQueue<T> {
             throw new NoSuchElementException("No such element for index " + index);
         }
         if (size == 1) {
-            assignEmptyQueue();
+            assignEmptyStack();
             return;
         }
         Object[] newData = new Object[size - 1];
@@ -53,17 +53,17 @@ public class MyArrayQueue<T> implements MyQueue<T> {
         if (size == 0) {
             return null;
         }
-        return (T) data[0];
+        return (T) data[data.length - 1];
     }
 
     @Override
     @SuppressWarnings("unchecked")
-    public T poll() {
+    public T pop() {
         if (size == 0) {
             return null;
         }
-        Object first = data[0];
-        remove(0);
+        Object first = data[data.length - 1];
+        remove(data.length - 1);
         return (T) first;
     }
 
@@ -76,7 +76,7 @@ public class MyArrayQueue<T> implements MyQueue<T> {
         return sj.toString();
     }
 
-    private void assignEmptyQueue() {
+    private void assignEmptyStack() {
         data = new Object[0];
         size = 0;
     }
